Enh: Pylint - R0801 for extinfo object. Remove useless method
