+++ test/install_files/install_root
@@ -309,6 +309,8 @@
 644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/contactgroup.pyc
 644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/escalation.py
 644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/escalation.pyc
+644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/genericextinfo.py
+644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/genericextinfo.pyc
 644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/host.py
 644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/host.pyc
 644 /usr/local/lib/PYTHONVERSION/dist-packages/alignak/objects/hostdependency.py
